+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 + Remove config folder (for old Docker)
 + Add Dockerfile
 + Add instructions for Docker
++ Add semaphore
 
 ## 4.2.0
 + Add hook-chaining support